api-phenotypic/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
autopep8==2.0.4
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
peewee==3.16.3
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
